Remove LoginActity by Fragment, create UserData singleton, remove some hardcoded string
