+++ app/src/main/java/treizieme/com/epinoish/LoginFragment.java
@@ -1,17 +1,42 @@
 package treizieme.com.epinoish;
 
 
+import android.app.ProgressDialog;
+import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
+import android.widget.EditText;
+import android.widget.Toast;
+
+import org.json.JSONException;
+import org.json.JSONObject;
+
+import java.io.IOException;
+
+import okhttp3.Callback;
+import okhttp3.FormBody;
+import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import okhttp3.RequestBody;
+import okhttp3.Response;
 
 
 /**
  * A simple {@link Fragment} subclass.
  */
 public class LoginFragment extends Fragment {
+
+    private String log;
+    private String pass;
+    private EditText login;
+    private EditText password;
+    private Button btnLogin;
+    private ProgressDialog progressDialog;
+    private final OkHttpClient client = new OkHttpClient();
 
 
     public LoginFragment() {
@@ -23,7 +48,125 @@ public class LoginFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_login, container, false);
+        View view = inflater.inflate(R.layout.fragment_login, container, false);
+
+        login = (EditText) view.findViewById(R.id.input_login);
+        password = (EditText) view.findViewById(R.id.input_password);
+
+        // Login btn listener
+        btnLogin = (Button) view.findViewById(R.id.btn_login);
+
+        btnLogin.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                login();
+            }
+        });
+
+        return view;
     }
 
+    private void login() {
+
+        if (loginValidate()) {
+
+            btnLogin.setEnabled(false);
+            progressDialog = new ProgressDialog(getActivity());
+            progressDialog.setIndeterminate(true);
+            progressDialog.setMessage("Login in progress...");
+            progressDialog.show();
+
+            RequestBody body = new FormBody.Builder()
+                    .add("login", log)
+                    .addEncoded("password", pass)
+                    .build();
+
+            Request request = new Request.Builder()
+                    .url("http://epitech-api.herokuapp.com/login")
+                    .post(body)
+                    .build();
+
+            client.newCall(request).enqueue(new Callback() {
+                @Override
+                public void onFailure(Request request, IOException e) {
+                    e.printStackTrace();
+                }
+
+                @Override
+                public void onResponse(final Response response) throws IOException {
+                    progressDialog.dismiss();
+
+                    if (response.code() == 401) {
+                        getActivity().runOnUiThread(new Runnable() {
+                            @Override
+                            public void run() {
+                                Toast.makeText(getActivity().getBaseContext(), "Bad login/password", Toast.LENGTH_LONG).show();
+                            }
+                        });
+                    } else if (!response.isSuccessful()) {
+                        getActivity().runOnUiThread(new Runnable() {
+                            @Override
+                            public void run() {
+                                Toast.makeText(getActivity().getBaseContext(), "Unexpected code " +
+                                                response.code() + " " + response.message(),
+                                        Toast.LENGTH_LONG).show();
+                            }
+                        });
+                        throw new IOException("Unexpected code " + response);
+                    } else {
+                        try {
+                            JSONObject json = new JSONObject(response.body().string());
+
+                            // Save token
+                            UserData userData = UserData.getInstance();
+                            userData.setLogin(log);
+                            userData.setPassword(pass);
+                            userData.setToken(json.getString("token"));
+
+                            SharedPreferences sharedPref = getActivity().getPreferences(0);
+                            SharedPreferences.Editor editor = sharedPref.edit();
+
+                            editor.putString("token", userData.getToken());
+                            editor.apply();
+
+                        } catch (JSONException e) {
+                            e.printStackTrace();
+                        }
+
+                        getActivity().runOnUiThread(new Runnable() {
+                            @Override
+                            public void run() {
+                                Toast.makeText(getActivity().getBaseContext(), "Login success", Toast.LENGTH_LONG).show();
+                                // End login fragment here call to profile view
+                            }
+                        });
+                    }
+                }
+            });
+        }
+        btnLogin.setEnabled(true);
+    }
+
+    private boolean loginValidate() {
+
+        log = login.getText().toString();
+        pass = password.getText().toString();
+        boolean isValid = true;
+
+        if (log.isEmpty()) {
+            login.setError("Please provide your login");
+            isValid = false;
+        } else {
+            login.setError(null);
+        }
+
+        if (pass.isEmpty()) {
+            password.setError("Please provide your password");
+            isValid = false;
+        } else {
+            password.setError(null);
+        }
+
+        return isValid;
+    }
 }
