+++ app/src/main/res/layout/content_login.xml
@@ -10,8 +10,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="treizieme.com.epinoish.LoginActivity"
-    tools:showIn="@layout/activity_login">
+    tools:context="treizieme.com.epinoish.LoginFragment">
 
     <!-- Login -->
     <android.support.design.widget.TextInputLayout
