+++ app/src/main/res/layout/nav_header_main.xml
@@ -18,6 +18,8 @@
         android:paddingTop="@dimen/nav_header_vertical_spacing"
         android:src="@android:drawable/sym_def_app_icon" />
 
+    <!-- TODO: REMOVE HARDCODED STRING! -->
+
     <TextView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
