+++ app/src/main/res/menu/activity_drawer_drawer.xml
@@ -3,14 +3,12 @@
 
     <group android:checkableBehavior="single">
         <item
-            android:id="@+id/nav_login_act"
-            android:icon="@drawable/ic_menu_camera"
-            android:title="Login by Activity" />
-        <item
             android:id="@+id/nav_login_frag"
             android:icon="@drawable/ic_menu_gallery"
-            android:title="Login by Fragment" />
+            android:title="@string/login" />
     </group>
+
+    <!-- TODO: REMOVE HARDCODED STRING HERE !!-->
 
     <item android:title="Communicate">
         <menu>
