+++ app/src/main/res/values/strings.xml
@@ -4,12 +4,10 @@
     <string name="input_hint_name">Name</string>
     <string name="input_hint_password">Password</string>
     <string name="login">Login</string>
+    <string name="title_activity_main">MainActivity</string>
     <string name="title_activity_login">LoginActivity</string>
     <string name="title_activity_drawer">DrawerActivity</string>
 
     <string name="navigation_drawer_open">Open navigation drawer</string>
     <string name="navigation_drawer_close">Close navigation drawer</string>
-
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
